+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.19.0
 * Deprecate `chargebackProtectionLevel` and add `protectionLevel` to `Dispute` and `DisputeSearchRequest`
 * Add `networkTokenizationAttributes` to `TransactionCreditCardRequest`
 * Add `NetworkTokenizationAttributesRequest` class